refactor: move cache.rs functions into util_cache.rs, decouple from OAuth2API impl for resuability.
