+++ src/oauth2/jwk/mod.rs
@@ -40,8 +40,8 @@
 
 pub mod jwk;
 
-mod cache;
 mod task;
 
 mod util;
+mod util_cache;
 mod util_refresh;
